Clean up message detail
